fix(arm-support): update simple auth bicep files
