+++ packages/fx-core/src/common/tools.ts
@@ -7,7 +7,6 @@ import {
   AzureSolutionSettings,
   ConfigFolderName,
   ConfigMap,
-  Dict,
   err,
   FxError,
   Json,
@@ -19,8 +18,6 @@ import {
   returnUserError,
   SolutionContext,
   SubscriptionInfo,
-  Tools,
-  UserError,
   UserInteraction,
 } from "@microsoft/teamsfx-api";
 import { promisify } from "util";
@@ -30,7 +27,6 @@ import * as path from "path";
 import * as uuid from "uuid";
 import { glob } from "glob";
 import { getResourceFolder } from "..";
-import { fakeServer } from "sinon";
 import { PluginNames } from "../plugins/solution/fx-solution/constants";
 import {
   AzureResourceApim,
@@ -42,6 +38,12 @@ import {
   MessageExtensionItem,
   TabOptionItem,
 } from "../plugins/solution/fx-solution/question";
+import * as Handlebars from "handlebars";
+
+Handlebars.registerHelper("contains", (value, array, options) => {
+  array = array instanceof Array ? array : [array];
+  return array.indexOf(value) > -1 ? options.fn(this) : "";
+});
 
 const execAsync = promisify(exec);
 
@@ -492,4 +494,25 @@ export function isFeatureFlagEnabled(featureFlagName: string): boolean {
     return true;
   }
   return false;
+}
+
+export function generateBicepFiles(
+  templateFilePath: string,
+  context: any
+): Result<string, FxError> {
+  try {
+    const templateString = fs.readFileSync(templateFilePath, "utf8");
+    const template = Handlebars.compile(templateString);
+
+    const updatedBicepFile = template(context);
+    return ok(updatedBicepFile);
+  } catch (error) {
+    return err(
+      returnSystemError(
+        new Error(`Failed to generate bicep file ${templateFilePath}. Reason: ${error.message}`),
+        "Core",
+        "BicepGenerationError"
+      )
+    );
+  }
 }
