+++ packages/fx-core/tests/plugins/resource/simpleauth/unit/expectedBicepFiles/onlyWithSimpleAuthPlugin/input_param.bicep
@@ -1,5 +1,5 @@
 
 param simpleAuth_sku string = 'F1'
-param simpleAuth_serverFarmsName string = '${resourceGroupName}-simpleAuth-serverfarms'
-param simpleAuth_webAppName string = '${resourceGroupName}-simpleAuth-webapp'
+param simpleAuth_serverFarmsName string = '${resourceBaseName}-simpleAuth-serverfarms'
+param simpleAuth_webAppName string = '${resourceBaseName}-simpleAuth-webapp'
