+++ packages/fx-core/tests/plugins/resource/simpleauth/unit/expectedBicepFiles/onlyWithSimpleAuthPlugin/simple_auth_test.bicep
@@ -1,18 +1,18 @@
 param sku string
 param simpleAuthServerFarmsName string
 param simpleAuthWebAppName string
-param tenantId string
+param m365TenantId string
 param AADClientId string
 @secure()
 param AADClientSecret string
 param applicationIdUri string
 
-param frontendHostingStorageEndpoint string
-
 var oauthAuthorityHost = environment().authentication.loginEndpoint
-var aadMetadataAddress = uri(oauthAuthorityHost, '${tenantId}/v2.0/.well-known/openid-configuration')
-var oauthAuthority = uri(oauthAuthorityHost, tenantId)
-var teamsAadIds = '1fec8e78-bce4-4aaf-ab1b-5451cc387264;5e3ce6c0-2b1f-4285-8d4b-75ee78787346'
+var aadMetadataAddress = uri(oauthAuthorityHost, '${m365TenantId}/v2.0/.well-known/openid-configuration')
+var oauthAuthority = uri(oauthAuthorityHost, m365TenantId)
+var teamsMobileOrDesktopAppClientId = '1fec8e78-bce4-4aaf-ab1b-5451cc387264'
+var teamsWebAppClientId = '5e3ce6c0-2b1f-4285-8d4b-75ee78787346'
+var authorizedClientApplicationIds = '${teamsMobileOrDesktopAppClientId};${teamsWebAppClientId}'
 
 resource simpleAuthServerFarms 'Microsoft.Web/serverfarms@2020-06-01' = {
   name: simpleAuthServerFarmsName
@@ -46,12 +46,11 @@ resource simpleAuthWebAppSettings 'Microsoft.Web/sites/config@2018-02-01' = {
   name: 'appsettings'
   properties: {
     AAD_METADATA_ADDRESS: aadMetadataAddress
-    ALLOWED_APP_IDS: teamsAadIds
+    ALLOWED_APP_IDS: authorizedClientApplicationIds
     IDENTIFIER_URI: applicationIdUri
     CLIENT_ID: AADClientId
     CLIENT_SECRET: AADClientSecret
     OAUTH_AUTHORITY: oauthAuthority
-    TAB_APP_ENDPOINT: frontendHostingStorageEndpoint
   }
 }
