+++ packages/fx-core/tests/plugins/resource/simpleauth/unit/expectedBicepFiles/withAllPlugins/module.bicep
@@ -1,5 +1,5 @@
 
-module simpleAuthProvision './simple_auth_test2.bicep' = {
+module simpleAuthProvision './simple_auth_test.bicep' = {
   name: 'simpleAuthProvision'
   params: {
     simpleAuthServerFarmsName: simpleAuth_serverFarmsName
@@ -9,6 +9,6 @@ module simpleAuthProvision './simple_auth_test2.bicep' = {
     AADClientSecret: AADClientSecret
     applicationIdUri: applicationIdUri
     frontendHostingStorageEndpoint: frontend_hosting_test_endpoint
-    tenantId: tenantId
+    m365TenantId: m365TenantId
   }
 }
